FRONTEND/requirements.txt
streamlit>=1.28.0
requests>=2.28.0
plotly>=5.15.0
pandas>=1.5.0
numpy>=1.21.0
rdkit-pypi>=2022.9.5
pillow>=9.0.0
python-dotenv>=1.0.0
